Read in all DB data on SeriesSet creation
It looks like there's no guarantee that a SeriesSet will be read until
the end, and if it's dropped early it will leak a database connection
potentially preventing future progress. This commit switches our builder
to read the entire result set at once so we can close the connection
immediately.

Additionally, keeping pgx.Rows around for long is dangerous as it can
induce a self-deadlock if the results from one query cause us to
perform another query.